update: some adjustments

+++ desktop/aur.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env bash
 
-sudo rm -rf /tmp/yay
-git clone https://aur.archlinux.org/yay-bin.git /tmp/yay
+if ! yay --version &>/dev/null; then
+	sudo rm -rf /tmp/yay
+	git clone https://aur.archlinux.org/yay-bin.git /tmp/yay
 
-makepkg -sri --noconfirm -D /tmp/yay
+	makepkg -sri --noconfirm -D /tmp/yay
+fi
